+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.1
+
+Released on Wednesday, August 5 2026.
+
+- Fixed `AmbiguousMatchException` in case of multiple `DomName` attributes (#136)
+- Added support for aliases with multiple `DomName` attributes
+
 # 1.0.0
 
 Released on Friday, July 31 2026.